requirements_pi.txt
asn1crypto==0.24.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.1.1
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpickle==1.2
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.1.0
pyasn1==0.4.7
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
python-telegram-bot==12.2.0
requests==2.22.0
six==1.12.0
tornado==6.0.3
urllib3==1.25.6
Werkzeug==0.15.6
